+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  - Fix inability to set camera device ([#519](https://github.com/kizniche/mycodo/issues/519))
  - Fix initialization of UART MHZ16 ([#520](https://github.com/kizniche/mycodo/issues/520))
+ - Fix issue with BMP280 ([#522](https://github.com/kizniche/mycodo/issues/522))
 
 ## 6.2.3 (2018-08-28)